refactor: extract `apiFetch` machinery as inner http client class
