+++ core/src/cloud/workflow-lifecycle.ts
@@ -12,7 +12,7 @@ import type { Log } from "../logger/log-entry.js"
 import { CloudApiError } from "../exceptions.js"
 import { gardenEnv } from "../constants.js"
 import type { Garden } from "../garden.js"
-import type { ApiFetchResponse } from "./api.js"
+import type { ApiFetchResponse } from "./http-client.js"
 import type { CreateWorkflowRunResponse } from "@garden-io/platform-api-types"
 import { dedent } from "../util/string.js"
 import { GotHttpError } from "../util/http.js"
